요구사항 보니까  파일을  두개라는 요구 사항이 있어서 파일을 두개로 분해 해서 다시 pr요청 합니다
